Refactor: SessionDto내로 SessionTrackerDto 이동
